feat: add trajectory recording decorator for chat completions
